Configured to capture spans for all evenEmitter listeners
