+++ src/loader.ts
@@ -24,7 +24,7 @@ const cli = sentryInit({
   dsn: process.env["SENTRY_DSN"],
   dist: process.env['COMMITHASH'],
   maxValueLength: 1000,
-  tracesSampleRate: 0,
+  tracesSampleRate: 1,
   sendDefaultPii: true,
   enableLogs: true,
   enableMetrics: true,
